Authentication Gardens

+++ urbangarden/gardens/views.py
@@ -10,7 +10,6 @@ from rest_framework import status
 from .models import Garden, Resource
 from .serializers import GardenSerializer, ResourceSerializer
 
-from rest_framework.views import APIView
 from rest_framework.decorators import api_view
 import coreapi
 from rest_framework.schemas import AutoSchema
@@ -55,13 +54,18 @@ class ResourceViewSchema(AutoSchema):
         return manual_fields + extra_fields
 
 
+from rest_framework.views import APIView
+import jwt
+from rest_framework.exceptions import AuthenticationFailed
+
+
 #Javier Martín - 04/12/2021
 
 # Create your views here.
 
 # GARDENS
-# 1 of 2: request for all ['GET', 'POST', 'DELETE']
-@api_view(['GET', 'POST', 'DELETE'])
+# 1 of 2: request for all ['GET']
+@api_view(['GET'])
 def garden_all(request):
     schema =GardenViewSchema()
     if request.method == 'GET':
@@ -69,34 +73,32 @@ def garden_all(request):
         garden_serializer = GardenSerializer(gardens, many=True)
         return JsonResponse(garden_serializer.data, safe=False)
         
-    elif request.method == 'POST':
-            garden_data = JSONParser().parse(request)
-            garden_serializer = GardenSerializer(data=garden_data)
-            if garden_serializer.is_valid():
-                garden_serializer.save()
-                return JsonResponse(garden_serializer.data, status=status.HTTP_201_CREATED) 
-            return JsonResponse(garden_serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-        
-    elif request.method == 'DELETE':
-        count = Garden.objects.all().delete()
-        return JsonResponse({'message': '{} Gardens were deleted successfully!'.format(count[0])}, status=status.HTTP_204_NO_CONTENT)
-
-
 # 2 of 2: request by garden ID ['GET', 'PUT', 'DELETE']
 
 @api_view(['GET', 'PUT', 'DELETE'])
 def garden_ID(request, pk):
-    schema =GardenViewSchema()
+    
+    token = request.COOKIES.get('jwt')
+    
     try: 
         garden = Garden.objects.get(pk=pk) 
     except Garden.DoesNotExist: 
         return JsonResponse({'message': 'The garden does not exist'}, status=status.HTTP_404_NOT_FOUND) 
  
     if request.method == 'GET': 
+
         garden_serializer = GardenSerializer(garden) 
         return JsonResponse(garden_serializer.data) 
  
     elif request.method == 'PUT': 
+
+        if not token:
+            raise AuthenticationFailed('Unauthenticated!')
+        try:
+            payload = jwt.decode(token, 'secret', algorithm=['HS256'])
+        except jwt.ExpiredSignatureError:
+            raise AuthenticationFailed('Unauthenticated!')
+
         garden_data = JSONParser().parse(request) 
         garden_serializer = GardenSerializer(garden, data=garden_data) 
         if garden_serializer.is_valid(): 
@@ -105,13 +107,21 @@ def garden_ID(request, pk):
         return JsonResponse(garden_serializer.errors, status=status.HTTP_400_BAD_REQUEST) 
  
     elif request.method == 'DELETE': 
+
+        if not token:
+            raise AuthenticationFailed('Unauthenticated!')
+        try:
+            payload = jwt.decode(token, 'secret', algorithm=['HS256'])
+        except jwt.ExpiredSignatureError:
+            raise AuthenticationFailed('Unauthenticated!')
+
         garden.delete() 
         return JsonResponse({'message': 'Garden was deleted successfully!'}, status=status.HTTP_204_NO_CONTENT)
 
 
 # RESOURCES
-# 1 of 2: request for all ['GET', 'POST', 'DELETE']
-@api_view(['GET', 'POST', 'DELETE'])
+# 1 of 2: request for all ['GET']
+@api_view(['GET'])
 def resource_all(request):
     schema =ResourceViewSchema()
     if request.method == 'GET':
@@ -119,18 +129,6 @@ def resource_all(request):
         resource_serializer = ResourceSerializer(resources, many=True)
         return JsonResponse(resource_serializer.data, safe=False)
         
-    elif request.method == 'POST':
-            resource_data = JSONParser().parse(request)
-            resource_serializer = ResourceSerializer(data=resource_data)
-            if resource_serializer.is_valid():
-                resource_serializer.save()
-                return JsonResponse(resource_serializer.data, status=status.HTTP_201_CREATED) 
-            return JsonResponse(resource_serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-        
-    elif request.method == 'DELETE':
-        count = Resource.objects.all().delete()
-        return JsonResponse({'message': '{} Resources were deleted successfully!'.format(count[0])}, status=status.HTTP_204_NO_CONTENT)
-
 
 # 2 of 2: request by resource ID ['GET', 'PUT', 'DELETE']
